+++ tests/dummy/app/serializers/application.js
@@ -1,0 +1,3 @@
+import ODataSerializer from 'ember-flexberry-data/serializers/odata';
+
+export default ODataSerializer.extend({});