+++ tests/dummy/app/serializers/employee-offline.js
@@ -1,0 +1,7 @@
+import { Serializer } from 'ember-flexberry-data';
+
+// TODO: ODataSerializer.extend
+export default Serializer.Offline.extend({
+  attrs: {
+  }
+});