+++ CHANGELOG.md
@@ -2,10 +2,55 @@
 
 ## [Unreleased]
 
+## [0.3.0] - 19.07.2016
+### Added
+- Support `polymorphic` options for `belongsTo` relationships.
+- Support special 'id' attribute at query language:
+
+  ```javascript
+  new SimplePredicate('manager', 'eq', key) === new SimplePredicate('manager.id', 'eq', key)
+  ```
+- Classes to support work in offline mode. See `Offline` namespace in [api documentation](http://flexberry.github.io/Documentation/master/modules/ember-flexberry-data.html).
+
+### Changed
+- The addon is no longer exports `Projection` namespace with following extra changes:
+    - `Projection`, `Offline`, `Adapter`, `Serializer`, `Query`, `Utils` namespaces are exporting now.
+    - `Projection.Store` mixin renamed to `Projection.StoreMixin`.
+    - `Projection.Adapter` mixin renamed to `Projection.AdapterMixin`.
+    - All adapters were included into `Adapter` napespace.
+    - All serializers were included into `Serializer` napespace.
+    - All client query language classes were included into `Query` napespace.
+    - `Information` class was included into `Utils` napespace.
+  For more information please see [api documentation](http://flexberry.github.io/Documentation/master/modules/ember-flexberry-data.html).
+
+### Fixed
+- Using `QueryBuilder.byId` method for OData query language adapter.
+
 ## [0.2.0] - 01.07.2016
 ### Added
 - Predicates for details with any / all functionality.
 - Filtering and ordering by master (using primary key) and it's fields.
 ### Fixed
-- Using nested complex predicates for all adapters.
+- Using nested complex predicates for all query language adapters.
 
+## [0.1.0] - 03.06.2016
+### Changed
+- Downgrade version of `broccoli-jscs` to 1.2.2.
+- Upgrade `ember-cli` to 2.4.3.
+
+## [0.0.3] - 02.06.2016
+### Added
+- Implementation of abstract query language adapter for IndexedDB.
+### Fixed
+- Logical operators in OData queries.
+
+## [0.0.2] - 25.05.2016
+### Fixed
+- Building OData URL for CRUD operations.
+
+## [0.0.1] - 20.05.2016
+The first release of ember-flexberry-data!
+### Added
+- Implementation of _projections_ - predefined named set of model attributes.
+- Implementation of abstract query language and adapters for OData and JSON.
+- Implementations of `DS.Store` and `DS.Adapter` with support of _projections_.